requirements.txt
# Binance API
python-binance>=1.0.19

# Environment
python-dotenv>=1.0.0

# Data & Backtest
pandas>=2.0.0
numpy>=1.24.0

# HTTP retry
requests>=2.28.0
